tests: lib: modem_trace: uart: delete test
The test used a strategy that leads to problems when introducing the
pinctrl: it is designed to run on qemu_cortex_m3 platform (not based on
nRF) but defining Nordic hardware on it. The test needs to improve code
isolation (ie testability) instead of making hacks.

Removing this test until an improved version is provided.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>