samples: matter: Unification of events, buttons and LEDs
This commit unifies events, buttons and LEDs for Matter
samples.

Changelog in this commit:

- Separated Events to the system-specific and application-specific.
- Created the TaskExecutor module which is responsible for posting and
dispatching application and system tasks (events).
- Created the common board configuration file (board_config.h)
that contains all board definitions, which can be overwritten in
the app_config.h file in the sample directory.
- Created the common board_const.h file that contains all constant
definitions of timeouts and LED blinking intervals.
Those definitions are the same for all samples and boards.
- Created the Board module in the board.c/.h files.
The Board module:
	- Initializes LEDs and buttons on the DK, taking
into account the number of LEDs and buttons on the DK.
	- Manages LEDs and buttons operations.
	- Manages to start Bluetooth LE Advertising using a button.
	- Exposes a button callback which can be used in the
Application code.
	- Manages the factory reset service.
- Aligned the Matter Lock sample to all changes.
	- Aligned the Lock operations after clicking the button.
	- Aligned the WiFi-Thread switching feature.
	- Aligned the SMP Bluetooth LE service DFU service.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>